+++ dime/lib/dime.dart
@@ -5,4 +5,5 @@ library dime;
 
 export 'src/common.dart';
 export 'src/dime_base.dart';
+export 'src/dime_static.dart';
 export 'src/factories.dart';