+++ dime/CHANGELOG.md
@@ -1,6 +1,15 @@
+## 0.3.0 - Dime static change
+
+- refactored to use static methods as `dime*` not `Dime.` static class
+- `Dime.` still available via deprecated API notation
+- changed name from `inject` to `get` to better match name to function of the method
+- `inject` still available via deprecated API notation
+- Code styles updates based on effective-dart lint rules.
+- Adding global scope property - to be used with Dime Flutter package (in development)
+
 ## 0.2.0 - Bug fix
 
-- added same module injection visible to next injections
+- added same module getion visible to next getions
 
 ## 0.1.1 - Scopes updates
 
@@ -9,9 +18,9 @@
 
 ## 0.1.0 - initial release
 
-- Dependency Injection with global scope.
-- Inject with modules and sub-scopes's modules.
+- Dependency getion with global scope.
+- get with modules and sub-scopes's modules.
 - Support for modules and scopes.
-- Creator types, on demand injection
+- Creator types, on demand getion
 - Ability to add singletons and singleton of same type per tag.
 - Closable types will be properly disposed when closing scopes